==> notebooks/2_data_cleaning_sephora.ipynb <==
# Sephora Data Cleaning

import pandas as pd
import sys
import os

# load Sephora dataset

sephora_df = pd.read_csv("../data/raw/sephora_df.csv")

sephora_df.head()

sephora_df[sephora_df['product_id']=='P467976']

sephora_df.head()

sephora_df[sephora_df['review_text'].notna()]

# For textual columns, fill NaNs with empty strings
text_columns = ['product_name', 'brand_name', 'primary_category', 'skin_type', 'review_text', 'ingredients']
sephora_df[text_columns] = sephora_df[text_columns].fillna('')

# For numerical columns, fill NaNs with zeros
numerical_columns = ['rating', 'is_recommended', 'reviews', 'price_usd', 'total_pos_feedback_count']
sephora_df[numerical_columns] = sephora_df[numerical_columns].fillna(0)

# clean Sephora cols with []
def clean_cell(x):
    if isinstance(x, float) and pd.isna(x):
        return ''
    if isinstance(x, str) and x.strip() == '[]':
        return ''
    if isinstance(x, list) and len(x) == 0:
        return ''
    return x

for col in text_columns:
    sephora_df[col] = sephora_df[col].apply(clean_cell)

sephora_df.head()

sephora_df.columns

sephora_df[sephora_df["product_id"] == "P467976"]

# my dataset has many rows of a similar products but 
# with different review texts, ratings, skintype
# the first row when search for a product id brign tota number of reviews, ingredients
# but the rest rows shows these as NAN
# FIXED

value = sephora_df.loc[
    (sephora_df["product_id"] == "P467976") & 
    (sephora_df["ingredients"].notnull()) &
    (sephora_df["ingredients"]!=''),
    "ingredients"
].iloc[0]

print(value)

# Clean Sephora DataFrame

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), '..')))

from utils.functions import clean_ingredient_input
sephora_df['ingredients'] = sephora_df['ingredients'].apply(clean_ingredient_input)

value = sephora_df.loc[
    (sephora_df["product_id"] == "P467976") & 
    (sephora_df["ingredients"].notnull()) &
    (sephora_df["ingredients"]!=''),
    "ingredients"
].iloc[0]

print(value)


sephora_df.dtypes

# aggregation
aggregation_functions = {
    'product_name': 'first',
    'brand_name': 'first',
    'ingredients': lambda x: list(set(ing for sublist in x.dropna() for ing in (sublist if isinstance(sublist, list) else [sublist]))),
    'rating': 'mean',
    'primary_category': 'first',
    'skin_type': lambda x: list(set(ty for sublist in x.dropna() for ty in (sublist if isinstance(sublist, list) else [sublist]))),
    'review_text': lambda x: list(x.dropna()),
    'is_recommended': 'mean',
    'reviews': 'first',
    'price_usd': 'mean',
    'total_pos_feedback_count': 'sum',
}

aggregated_df = sephora_df.groupby('product_id').agg(aggregation_functions).reset_index()

aggregated_df.head()

aggregated_df[aggregated_df['product_id']=='P467976']

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), '..')))

from utils.functions import clean_sephora_dataframe

aggregated_df = clean_sephora_dataframe(aggregated_df)




aggregated_df.head()

aggregated_df.tail()

aggregated_df.shape #(2286, 6)

# save data
aggregated_df.to_csv("../data/processed/sephora_df.csv", index=False)
