==> notebooks/3_data_cleaning_inci.ipynb <==
# INCI Data Cleaning

import pandas as pd
import sys
import os

# load inci dataset

inci_df = pd.read_csv("../data/raw/inci_df.csv")
inci_df.tail()

inci_df["who_is_it_good_for"].apply(type).value_counts()

# Map each ingredient to: What it does (e.g., “hydrating,” “acne-fighting”)
# Map each ingredient to: Which skin types it’s good or bad for (e.g., “Dry,” “Oily,” “Sensitive”)
# Clean dataset - null values, normalize text

sys.path.append(os.path.abspath(os.path.join(os.getcwd(), '..')))

from utils.functions import clean_inci_dataframe

inci_df = clean_inci_dataframe(inci_df)


inci_df.head(100)

inci_df.to_csv("../data/processed/inci_df.csv", index=False)
