==> notebooks/4_data_merge.ipynb <==
# Merge Sephora and INCI Dataframes based on their ingredients

import pandas as pd
import sys
import os

# load clean Sephora data
sephora_df = pd.read_csv("../data/processed/sephora_df.csv")

sephora_df.iloc[0].ingredients

sephora_df.iloc[0].product_id

#'P107306'

import ast
sephora_df['ingredients'] = sephora_df['ingredients'].apply(lambda x: ast.literal_eval(x) if isinstance(x, str) else x)

print(sephora_df['review_text'].apply(type).value_counts())


sephora_df.iloc[67].ingredients

df_exploded = sephora_df.explode('ingredients', ignore_index=True)


df_exploded.head()

df_exploded[df_exploded["product_id"]=="P107306"]

# clean text
sys.path.append(os.path.abspath(os.path.join(os.getcwd(), '..')))
from utils.functions import clean_review_list
import ast
import re

def safely_convert_string_to_list(val):
    try:
        if isinstance(val, str):
            result = ast.literal_eval(val)
            return result if isinstance(result, list) else []
        return []
    except:
        return []

def clean_text(text):
    pattern = r'[\\/*"\'`]'
    if isinstance(text, str) and text.strip():
        return re.sub(pattern, '', text).strip().lower()
    return ''

def clean_list_from_string(raw_value):
    text_list = safely_convert_string_to_list(raw_value)
    return [clean_text(text) for text in text_list if isinstance(text, str) and text.strip()]

df_exploded['review_text_cleaned'] = df_exploded['review_text'].apply(clean_list_from_string)

df_exploded.head()

# drop review text column
df_exploded.drop(columns=['review_text'], inplace=True)

df_exploded.head()

# clean skin_type column
df_exploded['skin_type'] = df_exploded['skin_type'].apply(clean_list_from_string)

df_exploded.head()

# load inci data
inci_df = pd.read_csv("../data/processed/inci_df.csv")

inci_df.head()

merged_df = pd.merge(df_exploded, inci_df, left_on='ingredients', right_on='name',
                     how='left')

merged_df.head()

# aggregation
aggregation_functions = {
    'product_name': 'first',
    'brand_name': 'first',
    'ingredients': lambda x: list(set(ing for sublist in x.dropna() for ing in (sublist if isinstance(sublist, list) else [sublist]))),
    'rating': 'mean',
    'primary_category': 'first',
    'skin_type': 'first',
    'review_text_cleaned': lambda x: list(x.dropna()),
    'is_recommended': 'mean',
    'reviews': 'first',
    'price_usd': 'mean',
    'total_pos_feedback_count': 'sum',
    
    # INCI fields to collect across ingredients:
    'what_does_it_do': lambda x: list(set(x.dropna())),
    'who_is_it_good_for': lambda x: list(set(x.dropna())),
    'who_should_avoid': lambda x: list(set(x.dropna())),
    'function_tags': lambda x: list(set(x.dropna())),
}

final_df = merged_df.groupby('product_id').agg(aggregation_functions).reset_index()

final_df.tail()

# save marged data
final_df.to_csv("../data/raw/merged_df.csv", index=False)

# Clean Dataframe Columns

import pandas as pd
import sys
import os

df = pd.read_csv("../data/raw/merged_df.csv")

df_copy = df

df_copy

# Clean following columns

# 'what_does_it_do',
# 'who_is_it_good_for', 
# 'who_should_avoid', 
# 'function_tags'

df_copy['what_does_it_do'] = df_copy['what_does_it_do'].apply(clean_list_from_string)

df_copy['who_is_it_good_for'] = df_copy['who_is_it_good_for'].apply(clean_list_from_string)

df_copy['who_should_avoid'] = df_copy['who_should_avoid'].apply(clean_list_from_string)

#function_tags
df_copy['function_tags'] = df_copy['function_tags'].apply(clean_list_from_string)

df_copy

type(final_df['review_text'].iloc[0])

# save *clean* marged data
df_copy.to_csv("../data/processed/merged_df.csv", index=False)
